requirements.txt
streamlit
pandas
PyMuPDF
langchain-core
langchain-together
langchain-community
